# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The musical mood/genre direction for the backing track.
 */
export type Vibe = typeof Vibe[keyof typeof Vibe];


export const Vibe = {
  lofi: 'lofi',
  cinematic: 'cinematic',
  pop: 'pop',
  rnb: 'rnb',
  electronic: 'electronic',
  acoustic: 'acoustic',
  ambient: 'ambient',
  serenity: 'serenity',
  soul: 'soul',
  jazz: 'jazz',
  folk: 'folk',
  afrobeat: 'afrobeat',
  synthwave: 'synthwave',
} as const;

/**
 * How long the generated song should be. "short" ≈ 30s, "standard" ≈ 90s, "long" ≈ 3 minutes. Longer songs take more time and credits to produce.
 */
export type SongLength = typeof SongLength[keyof typeof SongLength];


export const SongLength = {
  short: 'short',
  standard: 'standard',
  long: 'long',
} as const;

/**
 * Which engine produced the backing track. "elevenlabs" (the ElevenLabs Music model) is the only engine SynthScribe currently uses. "musicgen", "arranger", and "gpu" are legacy values kept only so older projects still deserialize; they are no longer offered.
 */
export type Engine = typeof Engine[keyof typeof Engine];


export const Engine = {
  musicgen: 'musicgen',
  elevenlabs: 'elevenlabs',
  arranger: 'arranger',
  gpu: 'gpu',
} as const;

/**
 * Engine used when creating a project. SynthScribe currently offers only ElevenLabs Music; legacy values ("musicgen", "arranger", "gpu") are not accepted for new projects.
 */
export type EngineChoice = typeof EngineChoice[keyof typeof EngineChoice];


export const EngineChoice = {
  elevenlabs: 'elevenlabs',
} as const;

/**
 * Which render path produces the lead vocal. "structural" (the default, "Option 1") lets ElevenLabs invent a melody in the hum's key/tempo/vibe while singing the lyrics. "note_for_note" ("Option 2") makes the lead vocal sing the lyrics on the EXACT hummed pitches and timing. Option 2 is currently selectable only for internal evaluation and falls back to Option 1 if its renderer is unavailable.
 */
export type RenderMode = typeof RenderMode[keyof typeof RenderMode];


export const RenderMode = {
  structural: 'structural',
  note_for_note: 'note_for_note',
} as const;

/**
 * Current stage of the hum-to-song pipeline.
 */
export type PipelineStage = typeof PipelineStage[keyof typeof PipelineStage];


export const PipelineStage = {
  draft: 'draft',
  transcribing: 'transcribing',
  generating_backing: 'generating_backing',
  singing: 'singing',
  mixing: 'mixing',
  complete: 'complete',
  error: 'error',
} as const;

export interface CreateProjectRequest {
  title?: string;
  vibe: Vibe;
  /** A short theme or story for AI-drafted lyrics. Stored for reference; used when drafting lyrics from the hum. */
  theme?: string;
  /** Lyrics the user wrote for the song. ElevenLabs sings these words over a backing in the chosen vibe. Optional — if omitted, an instrumental track is produced instead. */
  lyrics?: string;
  length?: SongLength;
  engine?: EngineChoice;
  renderMode?: RenderMode;
}

/**
 * Public URLs for each produced audio stage, when available.
 */
export interface AudioUrls {
  /** @nullable */
  hum: string | null;
  /** @nullable */
  backing: string | null;
  /** @nullable */
  vocals: string | null;
  /** @nullable */
  final: string | null;
}

export interface Project {
  id: string;
  title: string;
  vibe: Vibe;
  /**
     * The theme or story used to draft this song's lyrics, if any.
     * @nullable
     */
  theme?: string | null;
  /**
     * The lyrics ElevenLabs sang for this song, if any.
     * @nullable
     */
  lyrics?: string | null;
  length?: SongLength;
  engine: Engine;
  renderMode?: RenderMode;
  stage: PipelineStage;
  /** Coarse progress percentage (0-100). */
  progress: number;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  tempo?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  error?: string | null;
  audio: AudioUrls;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectStatus {
  id: string;
  stage: PipelineStage;
  progress: number;
  /**
     * Human-friendly status line, e.g. "Warming up the music model".
     * @nullable
     */
  message?: string | null;
  /** @nullable */
  error?: string | null;
  audio: AudioUrls;
}

/**
 * AI-drafted lyrics matched to a hummed melody.
 */
export interface DraftLyricsResponse {
  /** The drafted lyrics, one line per melodic phrase. */
  lyrics: string;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  tempo?: number | null;
  /** Number of lyric lines produced. */
  lineCount: number;
}

export interface ErrorResponse {
  error: string;
}

export type DraftLyricsBody = {
  file: Blob;
  /** A short theme or story the lyrics should be about. */
  theme: string;
  vibe?: Vibe;
};

export type UploadHumBody = {
  file: Blob;
};