# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SynthScribe API — turn a hummed melody into a fully produced song.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Song project operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List recent projects
      responses:
        "200":
          description: A list of projects, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new song project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by id
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: The project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lyrics/draft:
    post:
      operationId: draftLyrics
      tags: [projects]
      summary: Draft lyrics that match a hummed melody
      description: >-
        Analyzes an uploaded hum and uses an AI model to draft original lyrics
        whose line lengths and stresses are shaped to the melody's phrasing,
        with a mood matching the detected key/tempo and the chosen vibe. The
        returned lyrics are editable before producing a song.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                theme:
                  type: string
                  description: A short theme or story the lyrics should be about.
                vibe:
                  $ref: "#/components/schemas/Vibe"
              required:
                - file
                - theme
      responses:
        "200":
          description: Drafted lyrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftLyricsResponse"
        "400":
          description: Missing hum or theme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Lyric model unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/hum:
    post:
      operationId: uploadHum
      tags: [projects]
      summary: Upload the hummed recording for a project
      description: Accepts a 44.1kHz WAV recording of the user's hum.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Hum stored, project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/generate:
    post:
      operationId: startGeneration
      tags: [projects]
      summary: Start the hum-to-song pipeline
      description: Kicks off transcription, backing track, vocals, and mixing in the background.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "202":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No hum uploaded yet, or already generating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/status:
    get:
      operationId: getProjectStatus
      tags: [projects]
      summary: Poll the pipeline status for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Current pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Vibe:
      type: string
      description: The musical mood/genre direction for the backing track.
      enum:
        - lofi
        - cinematic
        - pop
        - rnb
        - electronic
        - acoustic
        - ambient
        - serenity
        - soul
        - jazz
        - folk
        - afrobeat
        - synthwave
    SongLength:
      type: string
      description: How long the generated song should be. "short" ≈ 30s, "standard" ≈ 90s, "long" ≈ 3 minutes. Longer songs take more time and credits to produce.
      enum:
        - short
        - standard
        - long
    Engine:
      type: string
      description: Which engine produced the backing track. "elevenlabs" (the ElevenLabs Music model) is the only engine SynthScribe currently uses. "musicgen", "arranger", and "gpu" are legacy values kept only so older projects still deserialize; they are no longer offered.
      enum:
        - musicgen
        - elevenlabs
        - arranger
        - gpu
    EngineChoice:
      type: string
      description: Engine used when creating a project. SynthScribe currently offers only ElevenLabs Music; legacy values ("musicgen", "arranger", "gpu") are not accepted for new projects.
      enum:
        - elevenlabs
    RenderMode:
      type: string
      description: >-
        Which render path produces the lead vocal. "structural" (the default,
        "Option 1") lets ElevenLabs invent a melody in the hum's key/tempo/vibe
        while singing the lyrics. "note_for_note" ("Option 2") makes the lead
        vocal sing the lyrics on the EXACT hummed pitches and timing. Option 2 is
        currently selectable only for internal evaluation and falls back to
        Option 1 if its renderer is unavailable.
      enum:
        - structural
        - note_for_note
    PipelineStage:
      type: string
      description: Current stage of the hum-to-song pipeline.
      enum:
        - draft
        - transcribing
        - generating_backing
        - singing
        - mixing
        - complete
        - error
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        vibe:
          $ref: "#/components/schemas/Vibe"
        theme:
          type: string
          description: A short theme or story for AI-drafted lyrics. Stored for reference; used when drafting lyrics from the hum.
        lyrics:
          type: string
          description: Lyrics the user wrote for the song. ElevenLabs sings these words over a backing in the chosen vibe. Optional — if omitted, an instrumental track is produced instead.
        length:
          $ref: "#/components/schemas/SongLength"
        engine:
          $ref: "#/components/schemas/EngineChoice"
        renderMode:
          $ref: "#/components/schemas/RenderMode"
      required:
        - vibe
    AudioUrls:
      type: object
      description: Public URLs for each produced audio stage, when available.
      properties:
        hum:
          type: [string, "null"]
        backing:
          type: [string, "null"]
        vocals:
          type: [string, "null"]
        final:
          type: [string, "null"]
      required:
        - hum
        - backing
        - vocals
        - final
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        vibe:
          $ref: "#/components/schemas/Vibe"
        theme:
          type: [string, "null"]
          description: The theme or story used to draft this song's lyrics, if any.
        lyrics:
          type: [string, "null"]
          description: The lyrics ElevenLabs sang for this song, if any.
        length:
          $ref: "#/components/schemas/SongLength"
        engine:
          $ref: "#/components/schemas/Engine"
        renderMode:
          $ref: "#/components/schemas/RenderMode"
        stage:
          $ref: "#/components/schemas/PipelineStage"
        progress:
          type: integer
          description: Coarse progress percentage (0-100).
        key:
          type: [string, "null"]
        tempo:
          type: [number, "null"]
        durationSeconds:
          type: [number, "null"]
        error:
          type: [string, "null"]
        audio:
          $ref: "#/components/schemas/AudioUrls"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - vibe
        - engine
        - stage
        - progress
        - audio
        - createdAt
        - updatedAt
    ProjectStatus:
      type: object
      properties:
        id:
          type: string
        stage:
          $ref: "#/components/schemas/PipelineStage"
        progress:
          type: integer
        message:
          type: [string, "null"]
          description: Human-friendly status line, e.g. "Warming up the music model".
        error:
          type: [string, "null"]
        audio:
          $ref: "#/components/schemas/AudioUrls"
      required:
        - id
        - stage
        - progress
        - audio
    DraftLyricsResponse:
      type: object
      description: AI-drafted lyrics matched to a hummed melody.
      properties:
        lyrics:
          type: string
          description: The drafted lyrics, one line per melodic phrase.
        key:
          type: [string, "null"]
        tempo:
          type: [number, "null"]
        lineCount:
          type: integer
          description: Number of lyric lines produced.
      required:
        - lyrics
        - lineCount
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error