# lib/api-zod/src/generated/types/createProjectRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */
import type { EngineChoice } from './engineChoice';
import type { RenderMode } from './renderMode';
import type { SongLength } from './songLength';
import type { Vibe } from './vibe';

export interface CreateProjectRequest {
  title?: string;
  vibe: Vibe;
  /** A short theme or story for AI-drafted lyrics. Stored for reference; used when drafting lyrics from the hum. */
  theme?: string;
  /** Lyrics the user wrote for the song. ElevenLabs sings these words over a backing in the chosen vibe. Optional — if omitted, an instrumental track is produced instead. */
  lyrics?: string;
  length?: SongLength;
  engine?: EngineChoice;
  renderMode?: RenderMode;
}