# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */
import type { AudioUrls } from './audioUrls';
import type { Engine } from './engine';
import type { PipelineStage } from './pipelineStage';
import type { RenderMode } from './renderMode';
import type { SongLength } from './songLength';
import type { Vibe } from './vibe';

export interface Project {
  id: string;
  title: string;
  vibe: Vibe;
  /**
     * The theme or story used to draft this song's lyrics, if any.
     * @nullable
     */
  theme?: string | null;
  /**
     * The lyrics ElevenLabs sang for this song, if any.
     * @nullable
     */
  lyrics?: string | null;
  length?: SongLength;
  engine: Engine;
  renderMode?: RenderMode;
  stage: PipelineStage;
  /** Coarse progress percentage (0-100). */
  progress: number;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  tempo?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  error?: string | null;
  audio: AudioUrls;
  createdAt: Date;
  updatedAt: Date;
}