# lib/api-zod/src/generated/types/renderMode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SynthScribe API — turn a hummed melody into a fully produced song.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Which render path produces the lead vocal. "structural" (the default, "Option 1") lets ElevenLabs invent a melody in the hum's key/tempo/vibe while singing the lyrics. "note_for_note" ("Option 2") makes the lead vocal sing the lyrics on the EXACT hummed pitches and timing. Option 2 is currently selectable only for internal evaluation and falls back to Option 1 if its renderer is unavailable.
 */
export type RenderMode = typeof RenderMode[keyof typeof RenderMode];


export const RenderMode = {
  structural: 'structural',
  note_for_note: 'note_for_note',
} as const;